now job can be deleted and successfully redirect

+++ src/components/modals/DeletePostJobsModal.js
@@ -1,21 +1,24 @@
-import React from 'react';
+import React, {useState} from 'react';
 import Modal from 'react-modal';
 import {FaArrowRight} from "react-icons/all";
 import "../../css/modals/post-job-delete-modal.css";
+import { Redirect } from "react-router-dom";
 
 const DeletePostJobsModal = (props) => {
-    const {isOpenModal, closeDeleteModal,jobID, removePostedJob} = props;
+    const {isOpenModal, closeDeleteModal,jobID, removePostedJob, isGoBack, GoBackTrue, goBackFromCards, goBackFromDetailsToJobHome} = props;
+
+    const [redirectTrue, setRedirectTrue] = useState(false);
+
 
     const handleDeletePostedJob = (jobID) => {
         removePostedJob(jobID);
         closeDeleteModal();
-        //if(GoBackTrue) isGoBack();
+        if(GoBackTrue){
+            if(goBackFromDetailsToJobHome) setRedirectTrue(true);
+            else isGoBack();
+        }
+        if(goBackFromCards) isGoBack();
     };
-
-
-
-
-
 
     const customStyles = {
         content : {
@@ -29,42 +32,48 @@ const DeletePostJobsModal = (props) => {
             border: 'none'
         }
     };
-    return (
-        <div className="container">
-            <Modal
-                isOpen = {isOpenModal}
-                style={customStyles}
-                ariaHideApp={false}
-                overlayClassName="Overlay"
-                closeTimeoutMS={300}
+    if(redirectTrue){
+        return (<Redirect to="/post-jobs"/>);
+    }
+    else {
 
-            >
-                <div className="row row-for-modals">
-                    <div className="col-12 col-for-modals">
-                        <span className="text-warning delete-modal-header">Delete</span>
-                        <i className="float-right text-light" onClick={() => closeDeleteModal()}><FaArrowRight/></i>
+        return (
+            <div className="container">
+                <Modal
+                    isOpen={isOpenModal}
+                    style={customStyles}
+                    ariaHideApp={false}
+                    overlayClassName="Overlay"
+                    closeTimeoutMS={300}
+
+                >
+                    <div className="row row-for-modals">
+                        <div className="col-12 col-for-modals">
+                            <span className="text-warning delete-modal-header">Delete</span>
+                            <i className="float-right text-light" onClick={() => closeDeleteModal()}><FaArrowRight/></i>
+                        </div>
                     </div>
-                </div>
-                <div className="row row-for-modals">
-                    <div className="col-12 col-for-modals">
-                        <p className="text-light">Are you sure to delete this Job?</p>
+                    <div className="row row-for-modals">
+                        <div className="col-12 col-for-modals">
+                            <p className="text-light">Are you sure to delete this Job?</p>
+                        </div>
                     </div>
-                </div>
 
-                <div className="row row-for-modals">
+                    <div className="row row-for-modals">
 
-                    <div className="col-12 col-for-modals">
-                        <button className="btn btn-danger" onClick={() => handleDeletePostedJob(jobID)}>YES</button>
-                        <button className="btn btn-info float-right" onClick={() => closeDeleteModal()}>NO</button>
+                        <div className="col-12 col-for-modals">
+                            <button className="btn btn-danger" onClick={() => handleDeletePostedJob(jobID)}>YES</button>
+                            <button className="btn btn-info float-right" onClick={() => closeDeleteModal()}>NO</button>
+                        </div>
                     </div>
-                </div>
 
 
-            </Modal>
+                </Modal>
 
-        </div>
+            </div>
 
-    );
+        );
+    }
 }
 
 export default DeletePostJobsModal;
